+++ changelog.md
@@ -2,16 +2,18 @@
 
 ## Changelog
 
-### Version 4.7 - 2024.01.09
+### Version 4.7 - 2024.01.15
 
 * New: option for expandable pairs has new layout and styling
 * New: expanded the `Elements` method `input` with more attributes
 * New: added several more UI icons
+* Edit: various improvements to the metabox handling and styling
 * Edit: settings `Render` class uses `Elements` for input types
 * Edit: various improvements for expandable pairs control
 * Edit: built-in cacert.pem 20231212
 * Fix: problem with the rendering of hierarchy checkboxes
 * Fix: missing default elements for expandable pairs control
+* Fix: several issues with validation of data for `Widget` class
 
 ### Version 4.6 - 2023.12.28